chore(lint): adopt oxlint-plugin-react-doctor and fix all findings
Add oxlint-plugin-react-doctor@0.9.12 (catalog-pinned) and enable 715 of
its native rules at error in .oxlintrc.json. Deliberate exclusions,
documented in the config:

- 117 ports of oxlint's built-in react/* and jsx-a11y/* rules stay off:
  66 duplicate built-ins already enforced with repo-specific options and
  per-file overrides, so violations report once, on the native rule.
- 5 preact/* rules stay off: they are not capability-gated and flag every
  `import { useRef } from 'react'` in this React app.
- 42 security-scan rules are no-ops under plain oxlint (CLI-only).
- Framework-gated buckets (nextjs, react-native, three/r3f, zustand,
  valtio) self-disable without their capability and are inert here.

Fixes for all 36 findings the new rules surfaced:

Bugs:
- invitations.accept.tsx: reset the accept spinner in `finally` (was
  stuck disabled on rejection); order loaderDeps before beforeLoad
  (loaderDeps after beforeLoad breaks type inference).
- blog.$slug.tsx, index.tsx: pass timeZone: 'UTC' to
  toLocaleDateString (server/browser date drift).
- cloudflare-workers-shim-dev.ts: parallelize independent dynamic
  imports with Promise.all.

Correctness/perf:
- sign-in.tsx, command-palette.tsx: replace post-mount
  useEffect(setState) flicker with useSyncExternalStore via new shared
  lib/client-only-value.ts (useHydrated, useClientValue).
- notification-feed.ts, migrations-fs.ts: single-pass loops instead of
  filter().map().
- Drop three manual useMemo view-model maps (pure derived data).
- command.tsx: duration-100 -> animation-duration-100 (tw-animate-css
  utility; duration-* read as transition: all) and gate animate-* with
  motion-safe: (also the sonner spinner).
- Skip links apply metrics unconditionally so focus does not shift
  layout; remove no-op space-y-0 on flex CardHeaders.
- Copy: em dashes and three-dot ellipses replaced.

Structure:
- Split the 350-line HomePage into one-component-per-file sections
  under components/landing/ (no-giant-component, no-multi-component-file).

Documented disable comments (with justification): prefer-dynamic-import
on chart files (TanStack Start autoCodeSplitting already route-chunks
recharts; mdx-chart is React.lazy-consumed), anchor-target-exists on
the layout skip link (target owned by route children), effect rules at
the platform boundary and in the event-handler finally.

lint: 0 errors on 1187 rules; typecheck, tests, build, format all green.
