+++ apps/web/src/components/landing/github-url.ts
@@ -1,0 +1,3 @@
+/** The repository this starter showcases; shared by the hero and closing sections. */
+export const GITHUB_URL =
+  'https://github.com/brandhaug/full-stack-typescript-monorepo-starter-with-authentication'
