+++ apps/web/src/components/ui/sonner.tsx
@@ -13,7 +13,7 @@ const TOASTER_ICONS = {
   info: <InfoIcon className="size-4" />,
   warning: <TriangleAlertIcon className="size-4" />,
   error: <OctagonXIcon className="size-4" />,
-  loading: <Loader2Icon className="size-4 animate-spin" />
+  loading: <Loader2Icon className="size-4 motion-safe:animate-spin" />
 }
 
 // Sonner is themed through CSS custom properties, which `CSSProperties` alone
