+++ apps/web/src/components/workspace-shell.tsx
@@ -71,7 +71,7 @@ export function WorkspaceShell({
     <div className="grid min-h-dvh bg-background lg:grid-cols-[16rem_1fr]">
       <a
         href="#main-content"
-        className="sr-only focus:not-sr-only focus:fixed focus:top-2 focus:left-2 focus:z-50 focus:rounded-md focus:bg-primary focus:px-3 focus:py-2 focus:text-sm focus:text-primary-foreground"
+        className="sr-only focus:not-sr-only focus:fixed focus:top-2 focus:left-2 focus:z-50 focus:rounded-md focus:bg-primary px-3 py-2 text-sm focus:text-primary-foreground"
       >
         Skip to content
       </a>
