+++ packages/db/src/migrations-fs.ts
@@ -13,12 +13,15 @@ const migrationsDir = join(import.meta.dirname, '..', 'migrations')
 
 /** Reads every committed migration, sorted by folder name (timestamp order). */
 export function listMigrations(): Array<{ name: string; sql: string }> {
-  return readdirSync(migrationsDir, { withFileTypes: true })
-    .filter((entry) => entry.isDirectory())
-    .map((entry) => entry.name)
-    .toSorted()
-    .map((name) => ({
-      name,
-      sql: readFileSync(join(migrationsDir, name, 'migration.sql'), 'utf8')
-    }))
+  const names = readdirSync(migrationsDir, { withFileTypes: true }).reduce<string[]>(
+    (folders, entry) => {
+      if (entry.isDirectory()) folders.push(entry.name)
+      return folders
+    },
+    []
+  )
+  return names.toSorted().map((name) => ({
+    name,
+    sql: readFileSync(join(migrationsDir, name, 'migration.sql'), 'utf8')
+  }))
 }
